Fix type of column cs_sold_date_sk (to match type of d_date_sk)

+++ scripts/impala-parquet.sql
@@ -488,7 +488,7 @@ create table catalog_sales
   cs_net_paid_inc_ship_tax  decimal(7,2),
   cs_net_profit             decimal(7,2)
 )
-partitioned by (cs_sold_date_sk bigint)
+partitioned by (cs_sold_date_sk integer)
 stored as parquet;
 
 create table web_returns
